+++ frontend/src/panel/api/scripts.ts
